refactor(stack): extract data, types, and grid into separate modules
